Combine SurveyTaskContainer and SurveyTask components and tests
