+++ packages/lib-classifier/src/plugins/tasks/survey/components/SurveyTask.js
@@ -1,24 +1,91 @@
 import { Box } from 'grommet'
+import { observer } from 'mobx-react'
 import PropTypes from 'prop-types'
-import React from 'react'
+import React, { useEffect } from 'react'
 
 import Choice from './components/Choice'
 import Chooser from './components/Chooser'
 
-function SurveyTask({
-  answers = {},
+export function SurveyTaskContainer ({
+  annotation,
+  task,
   disabled = false,
-  filters = {},
-  previousChoice = '',
-  handleAnswers = () => {},
-  handleChoice = () => {},
-  handleDelete = () => {},
-  handleFilter = () => {},
-  handleIdentify = () => {},
-  selectedChoice = '',
-  selectedChoiceIds = [],
-  task
 }) {
+  const [answers, setAnswers] = React.useState({})
+  const [filters, setFilters] = React.useState({})
+  const [previousChoice, setPreviousChoice] = React.useState('')
+  const [selectedChoice, setSelectedChoice] = React.useState('')
+
+  useEffect(() => {
+    return () => {
+      if (annotation._choiceInProgress) {
+        annotation.setChoiceInProgress(false)
+      }
+    }
+  })
+
+  function handleAnswers (answers) {
+    setAnswers(answers)
+  }
+
+  function handleChoice (selectedChoice) {
+    if (selectedChoice === '') {
+      annotation.setChoiceInProgress(false)
+    } else {
+      annotation.setChoiceInProgress(true)
+    }
+
+    if (annotation?.value?.map(item => item.choice).includes(selectedChoice)) {
+      const existingAnnotationValue = annotation?.value?.find(value => value.choice === selectedChoice)
+      setAnswers(existingAnnotationValue.answers)
+      setSelectedChoice(selectedChoice)
+    } else {
+      setAnswers({})
+      setSelectedChoice(selectedChoice)
+    }
+  }
+
+  function handleDelete (deletedChoice) {
+    if (annotation?.value?.map(item => item.choice).includes(deletedChoice)) {
+      const value = annotation?.value?.filter(item => item.choice !== deletedChoice) || []
+      annotation.update(value)
+    }
+    
+    setAnswers({})
+    setPreviousChoice(deletedChoice)
+    setSelectedChoice('')
+    
+    annotation.setChoiceInProgress(false)
+  }
+
+  function handleFilter (characteristicId, valueId) {
+    let newFilters = Object.assign({}, filters)
+    if (valueId) {
+      newFilters[characteristicId] = valueId
+    } else if (characteristicId) {
+      delete newFilters[characteristicId]
+    } else {
+      newFilters = {}
+    }
+
+    setFilters(newFilters)
+  }
+
+  function handleIdentify () {
+    const parsedFilters = JSON.parse(JSON.stringify(filters))
+    const value = annotation?.value?.filter(item => item.choice !== selectedChoice) || []
+    value.push({ choice: selectedChoice, answers, filters: parsedFilters })
+
+    setAnswers({})
+    setSelectedChoice('')
+    setPreviousChoice(selectedChoice)
+
+    annotation.update(value)
+    annotation.setChoiceInProgress(false)
+  }
+
+  const selectedChoiceIds = annotation?.value?.map(item => item.choice)
+
   return (
     <Box
       fill
@@ -47,23 +114,14 @@ function SurveyTask({
   )
 }
 
-SurveyTask.propTypes = {
-  answers: PropTypes.objectOf(
-    PropTypes.oneOfType([
-      PropTypes.arrayOf(PropTypes.string),
-      PropTypes.string
-    ])
-  ),
+SurveyTaskContainer.propTypes = {
+  annotation: PropTypes.shape({
+    setChoiceInProgress: PropTypes.func,
+    update: PropTypes.func,
+    value: PropTypes.array,
+    _choiceInProgress: PropTypes.bool
+  }).isRequired,
   disabled: PropTypes.bool,
-  filters: PropTypes.objectOf(PropTypes.string),
-  previousChoice: PropTypes.string,
-  handleAnswers: PropTypes.func,
-  handleChoice: PropTypes.func,
-  handleDelete: PropTypes.func,
-  handleFilter: PropTypes.func,
-  handleIdentify: PropTypes.func,
-  selectedChoice: PropTypes.string,
-  selectedChoiceIds: PropTypes.arrayOf(PropTypes.string),
   task: PropTypes.shape({
     help: PropTypes.string,
     required: PropTypes.oneOfType([PropTypes.string, PropTypes.bool]),
@@ -72,4 +130,4 @@ SurveyTask.propTypes = {
   }).isRequired
 }
 
-export default SurveyTask
+export default observer(SurveyTaskContainer)
