+++ packages/lib-classifier/src/plugins/tasks/survey/components/index.js
@@ -1,2 +1,1 @@
-export { default } from './SurveyTaskContainer'
-export { default as SurveyTask } from './SurveyTask'
+export { default } from './SurveyTask'
